Add from_email and reply_to attrs to EmailTemplate
This allows us to manage these properties from the admin site. If you pass a value
through to `create_message` it will take precedence, but in the absence of one, they
will default to what is stored in the database. This is good for standard emails, and
the defaults for these defaults is DEFAULT_FROM_EMAIL, which in turn defaults
to `webmaster@localhost`. Is that enough defaults for you?