Misc: Fix slow loading LoggedMessage admin (#32)
Previously, the `list_filter` parameter for `LoggedMessage` contained `template__name` and `template__language` which were designed to allow us filtering of the logged message results set by listing all template names and languages used in the original set.

The problem with this is that it requires a join with `EmailTemplate` and over a large result set this can become extremely slow.

The reality is that we just need a way to filter based on all known templates & languages, rather than those specific to the exact set we're looking at; and thus the solution to this is to implement two custom `SimpleListFilter`s to do that job. These use simple non-join selects to retrieve the data instead.

This PR also makes a design change to the template name filter itself within the `LoggedMessage` change list page: it now displays the names at a smaller font-size so more text can be displayed on a single line and ensures they do not split onto multiple lines which hurt readability of the filter. Instead, they now curtail with an ellipsis at a certain width and the user can instead see the full value by hovering over the link thanks to the new `title` attributes added.